% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_files}
\alias{fetch_files}
\title{Retrieve FCS files from ImmPort}
\usage{
fetch_files(study, output_dir = ".")
}
\arguments{
\item{study}{A character. ImmPort study accession.}

\item{output_dir}{A character. Output directory.}
}
\value{
A character. Output path.
}
\description{
Retrieve FCS files from ImmPort
}
\examples{
\dontrun{
fetch_files("SDY820")
}

}
