% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_study}
\alias{process_study}
\title{Pre-process a ImmPort study with flow cytometry data}
\usage{
process_study(study, input_dir, debug_dir = NULL)
}
\arguments{
\item{study}{A character. ImmPort study accession.}

\item{input_dir}{A character. Input directory.}

\item{debug_dir}{A character. Debug directory.}
}
\value{
A list of gating set objects.
}
\description{
Pre-process a ImmPort study with flow cytometry data
}
\examples{
\dontrun{
gsl <- process_study("SDY820", "SDY820/ResultFiles/Flow_cytometry_result")
}

}
