% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_gating_sets}
\alias{save_gating_sets}
\title{Save gating sets and create QC reports and plots}
\usage{
save_gating_sets(gsl, output_dir, qc = TRUE)
}
\arguments{
\item{gsl}{A list. List of gating sets to save.}

\item{output_dir}{A character. Output directory.}

\item{qc}{A logical. Create QC reports/plots.}
}
\value{
A character. Output path.
}
\description{
Save gating sets and create QC reports and plots
}
\examples{
\dontrun{
save_gating_sets(gsl, "SDY820/gs_dir")
}

}
